feat(abi): wire manifest rate-limit through WasmSupervisor::call_verb
Resolves BACKLOG "Rate-limit declarations are unenforced". Before this
commit the manifest's `[rate_limit]` block reached the SDK `RateLimit`
struct but the token bucket was only constructed on the unused
`PluginSupervisor`; every real verb call ran un-throttled.

Three coupled fixes needed for the feature to be real:

1. **Manifest location.** Every bundled plugin put the block under
   `[permissions.rate_limit]`, but `PluginManifest.rate_limit` is a
   top-level field, so the section was silently absorbed into
   `Permissions._extra` and never reached the supervisor. All 7
   plugin.tomls now use top-level `[rate_limit]`.

2. **Field name + type.** `RateLimit.rps: u32` missed the `rps` ↔
   `requests_per_second` rename the plugin.tomls had been using, and
   even if the name matched `u32` would silently clamp fractional
   values (`0.4` → `0`) — disabling every sub-unit rate. Widened to
   `f32` with `#[serde(rename = "requests_per_second")]`. `TokenBucket::
   new` now takes `f32` throughout.

3. **Wire-through.** `WasmSupervisor` now owns its own
   `Option<TokenBucket>`, built from the manifest block at `load()`
   time. `call_verb` awaits a token BEFORE starting the call-timeout
   clock so queue delay doesn't cannibalise the execution budget.
   Plugins without a rate_limit block skip the acquire.

Two new `#[tokio::test(start_paused = true)]` cases lock in the
fractional-rps path (discogs 0.4 rps: burst 5 drains instantly, 6th
acquire waits 2.5s) and the disabled-bucket case (`rps <= 0`).

Live smoke: all 7 plugins still load to the expected state, zero
reload events, first call after load is as fast as before (burst 5–10
covers the catalog-startup fanout).

+++ sdk/src/manifest.rs
@@ -426,7 +426,11 @@ pub struct AuthorMeta {
 #[derive(Debug, Clone, Deserialize, Serialize)]
 pub struct RateLimit {
     /// Steady-state requests per second (tokens generated per second).
-    pub rps: u32,
+    ///
+    /// Fractional values are supported and meaningful: `0.4` = 24 req/min
+    /// (Discogs unauthenticated tier), `0.1` = 6 req/min (libre.fm).
+    #[serde(rename = "requests_per_second")]
+    pub rps: f32,
     /// Maximum burst capacity — the bucket size.
     #[serde(default = "default_burst")]
     pub burst: u32,
